requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.3.1
django-multiselectfield==0.1.12
django-rest==0.8.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
gunicorn==21.2.0
pyyaml==6.0.1
uritemplate==4.1.1
mailtrap==2.0.1